+++ Krowi_AchievementFilter/ExpansionData/2 The Burning Crusade/Raids.lua
@@ -49,7 +49,7 @@ end
 raids.ZulAman = {}; -- 2.3
 function raids.ZulAman.Load(catRaids)
     local raid = catRaids:AddCatInstanceInfo(77);
-    raid.Name = raid.Name .. " (" .. GetCategoryInfo(15234) .. ")";
+    raid.Name = raid.Name .. " (" .. addon.GetCategoryInfoTitle(15234) .. ")";
     raid:AddAchievementIDsUnobtainable(691); -- Defeat the bosses in X (Unobtainable)
     raid:AddAchievementIDsUnobtainable(430); -- Mounts (Unobtainable)
 end