+++ Krowi_AchievementFilter/ExpansionData/3 Wrath of the Lich King/Raids.lua
@@ -73,7 +73,7 @@ function raids.Ulduar.Load(catRaids)
     raid:AddAchievementIDs(12297, 12302, 12309, 12310, 12311, 12399); -- Defeat the bosses in X
     
     local version = raid:AddCatDifficulty10();
-    version.Name = version.Name .. " (" .. GetCategoryInfo(15234) .. ")";
+    version.Name = version.Name .. " (" .. addon.GetCategoryInfoTitle(15234) .. ")";
     version:AddAchievementIDsUnobtainable(3097, 2907, 2905, 2911, 2909, 2913, 2914, 2915, 3056, -- Flame Leviathan
                                             2925, 2927, 2930, -- Ignis the Furnace
                                             2919, 2923, -- Razorscale
@@ -92,7 +92,7 @@ function raids.Ulduar.Load(catRaids)
     version:AddAchievementIDsUnobtainable(2886, 2888, 2890, 2892, 2894, 3036); -- Defeat the bosses in X
 
     version = raid:AddCatDifficulty25();
-    version.Name = version.Name .. " (" .. GetCategoryInfo(15234) .. ")";
+    version.Name = version.Name .. " (" .. addon.GetCategoryInfoTitle(15234) .. ")";
     version:AddAchievementIDsUnobtainable(3098, 2908, 2906, 2912, 2910, 2918, 2916, 2917, 3057, -- Flame Leviathan
                                             2926, 2928, 2929, -- Ignis the Furnace
                                             2921, 2924, -- Razorscale