+++ Krowi_AchievementFilter/Krowi_AchievementFilter.lua
@@ -23,10 +23,10 @@ function loadHelper:OnEvent(event, arg1)
             addon.Icon.Load();
             addon.Tutorials.Load();
         elseif arg1 == "Blizzard_AchievementUI" then -- This needs the Blizzard_AchievementUI addon available to load
-            addon.GUI.SearchBox1 = addon.GUI.SearchBox:New();
+            -- addon.GUI.SearchBox1 = addon.GUI.SearchBox:New();
             addon.GUI.AchievementsFrame1 = addon.GUI.AchievementsFrame:New();
             addon.GUI.CategoriesFrame1 = addon.GUI.CategoriesFrame:New(addon.Categories, addon.GUI.AchievementsFrame1.Frame);
-            addon.GUI.TabButton1 = addon.GUI.AchievementFrameTabButton:New(addon.L["T_TAB_TEXT"], addon.GUI.CategoriesFrame1.Frame, addon.GUI.AchievementsFrame1.Frame, addon.GUI.SearchBox1.Frame);
+            addon.GUI.TabButton1 = addon.GUI.AchievementFrameTabButton:New(addon.L["T_TAB_TEXT"], addon.GUI.CategoriesFrame1.Frame, addon.GUI.AchievementsFrame1.Frame); -- addon.GUI.SearchBox1.Frame
             addon.Tutorials.HookTrigger(addon.GUI.TabButton1);
         end
     elseif event == "PLAYER_LOGIN" then -- This needs player achievement info which is not yet available on "ADDON_LOADED"
@@ -40,6 +40,43 @@ function loadHelper:OnEvent(event, arg1)
         addon.BattleForAzeroth.Load();
         addon.Shadowlands.Load();
         addon.Categories, addon.Achievements = addon.Data:GetList(); -- This freezes the game if we load it outside of the loading screen
+
+        TEST = {};
+        if addon.Diagnostics.DebugEnabled() then
+            for _, category in next, addon.Categories do
+                local parentName = "";
+                if category.Parent ~= nil then
+                    parentName = category.Parent.Name;
+                end
+                tinsert(TEST, {parentName, category.Name});
+            end
+        end
     end
 end
 loadHelper:SetScript("OnEvent", loadHelper.OnEvent);
+
+function addon.ResetView(categoriesButtons)
+    addon.Diagnostics.Trace("addon.ResetView");
+    
+    KrowiAF_AchievementCategoryButton_OnClick(categoriesButtons[1]); -- Select the 1st category
+    if categoriesButtons[1].Category.NotCollapsed then -- Make sure it's collapsed
+        KrowiAF_AchievementCategoryButton_OnClick(categoriesButtons[1]);
+    end
+end
+
+function addon.GetCategoryInfoTitle(categoryID)
+    addon.Diagnostics.Trace("addon.GetCategoryInfoTitle");
+
+    local title = GetCategoryInfo(categoryID);
+    return title;
+end
+
+function addon.GetAchievement(id)
+    addon.Diagnostics.Trace("addon.GetAchievement");
+
+	for _, achievement in next, addon.Achievements do
+		if achievement.ID == id then
+			return achievement;
+		end
+	end
+end