+++ Krowi_AchievementFilter/Objects/Achievement.lua
@@ -22,3 +22,17 @@ end
 function achievement:AddRCMenExtra(rcMenExtra)
     self.RCMenExtra = rcMenExtra;
 end
+
+function achievement:GetCategory()
+    addon.Diagnostics.Trace("achievement:GetCategory");
+
+    for _, category in next, addon.Categories do
+        if category.Achievements ~= nil then
+            for _, achievement in next, category.Achievements do
+                if achievement.ID == self.ID then
+                    return category;
+                end
+            end
+        end
+    end
+end