+++ Krowi_AchievementFilter/CHANGELOG.md
@@ -4,7 +4,10 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 ### Added
 - Searching achievements in the new tab
-- Watch achievements
+
+### Changed
+- Moved some code build frames into Templates.xml
+- Refactored functions to select categories and achievements (needed for searching achievements)
 
 ## [9.0.2.8.1] - 2021-01-16
 ### Fixed